+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestOnDemandStatus.java
@@ -18,7 +18,7 @@ import com.datastax.oss.driver.api.core.cql.ResultSet;
 import com.datastax.oss.driver.api.core.cql.Row;
 import com.datastax.oss.driver.api.core.data.UdtValue;
 import com.datastax.oss.driver.api.core.type.UserDefinedType;
-import com.ericsson.bss.cassandra.ecchronos.core.AbstractCassandraTest;
+import com.ericsson.bss.cassandra.ecchronos.core.AbstractCassandraContainerTest;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.OngoingJob.Status;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.ReplicationState;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
@@ -46,7 +46,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.Silent.class)
-public class TestOnDemandStatus extends AbstractCassandraTest
+public class TestOnDemandStatus extends AbstractCassandraContainerTest
 {
     private static final String STATUS_FAILED = "failed";
     private static final String STATUS_FINISHED = "finished";